test: disable workspace upgrade step for tests

+++ packages/nocodb/src/ee/modules/jobs/jobs/source-create/source-create.processor.ts
@@ -53,28 +53,36 @@ export class SourceCreateProcessor {
 
     logBasic('🚀 Your data source schema is loaded successfully!');
 
-    if (needUpgrade) {
-      logBasic(' ');
-      logBasic(
-        '***************************************************************************',
-      );
-      logBasic('🌟 Setting up a dedicated instance for your first datasource.');
-      logBasic(
-        '⏳ This may take 3-5 minutes, so perfect time to grab a cup of coffee and we would be all set.',
-      );
-      logBasic('Thank you for your patience 🙏');
-      logBasic('For next data sources in your workspace this will be instant.');
-      logBasic(
-        '***************************************************************************',
-      );
-      logBasic(' ');
+    if (process.env.NC_TEST_EE !== 'true') {
+      if (needUpgrade) {
+        logBasic(' ');
+        logBasic(
+          '***************************************************************************',
+        );
+        logBasic(
+          '🌟 Setting up a dedicated instance for your first datasource.',
+        );
+        logBasic(
+          '⏳ This may take 3-5 minutes, so perfect time to grab a cup of coffee and we would be all set.',
+        );
+        logBasic('Thank you for your patience 🙏');
+        logBasic(
+          'For next data sources in your workspace this will be instant.',
+        );
+        logBasic(
+          '***************************************************************************',
+        );
+        logBasic(' ');
 
-      await NcConnectionMgrv2.deleteAwait(createdBase);
+        await NcConnectionMgrv2.deleteAwait(createdBase);
 
-      await this.workspacesService.upgrade({
-        workspaceId,
-        user,
-      });
+        await this.workspacesService.upgrade({
+          workspaceId,
+          user,
+        });
+      }
+    } else {
+      needUpgrade = false;
     }
 
     this.debugLog(`job completed for ${job.id}`);
